Move notifyNewMember out of the use-server notifications module
src/notifications.ts is "use server", so exporting notifyNewMember there
made it a wire-callable action anyone could invoke with arbitrary
publication/membership ids. It now lives in src/membership.server.ts, a
plain server module only importable by server code.
